pyproject.toml
[project]
name = "drawio2pptx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autopep8>=2.3.2",
    "defusedxml>=0.7.1",
    "flask>=3.1.2",
    "lxml>=6.0.2",
    "python-pptx>=1.0.2",
]
